Move code that creates Signature to astbuilder module
