+++ pydoctor/model.py
@@ -15,6 +15,7 @@ import platform
 import sys
 import types
 from enum import Enum
+from inspect import Signature
 from optparse import Values
 from pathlib import Path
 from typing import (
@@ -456,6 +457,7 @@ class Function(Inheritable):
     annotations: Mapping[str, Optional[ast.expr]]
     decorators: Optional[Sequence[ast.expr]]
     argspec: Tuple[Sequence[str], Optional[str], Optional[str], Sequence[str]]
+    signature: Signature
 
     def setup(self) -> None:
         super().setup()
@@ -710,6 +712,7 @@ class System:
                 f.docstring = v.__doc__
                 f.decorators = None
                 f.argspec = ((), None, None, ())
+                f.signature = Signature()
                 self.addObject(f)
             elif isinstance(v, type):
                 c = self.Class(self, k, parent)
